+++ tests/CRUDlexTests/CRUDEntityTest.php
@@ -125,6 +125,11 @@ class CRUDEntityTest extends \PHPUnit_Framework_TestCase {
                     'required' => false,
                     'unique' => false,
                     'input' => false
+                ),
+                'isOpenOnSundays' => array(
+                    'required' => false,
+                    'unique' => false,
+                    'input' => false
                 )
             )
         );