+++ tests/CRUDlexTests/CRUDMySQLDataTest.php
@@ -232,4 +232,37 @@ class CRUDMySQLDataTest extends \PHPUnit_Framework_TestCase {
         $this->dataBook->fetchReferences(null);
     }
 
+    public function testBoolHandling() {
+        $libraryA = $this->dataLibrary->createEmpty();
+        $libraryA->set('name', 'lib');
+        $this->dataLibrary->create($libraryA);
+
+        $read = $this->dataLibrary->get($libraryA->get('id'))->get('isOpenOnSundays');
+        $expected = '0';
+        $this->assertSame($read, $expected);
+
+        $libraryB = $this->dataLibrary->createEmpty();
+        $libraryB->set('name', 'lib');
+        $libraryB->set('isOpenOnSundays', '1');
+        $this->dataLibrary->create($libraryB);
+
+        $read = $this->dataLibrary->get($libraryB->get('id'))->get('isOpenOnSundays');
+        $expected = '1';
+        $this->assertSame($read, $expected);
+
+        $libraryA->set('isOpenOnSundays', '1');
+        $this->dataLibrary->update($libraryA);
+
+        $read = $this->dataLibrary->get($libraryA->get('id'))->get('isOpenOnSundays');
+        $expected = '1';
+        $this->assertSame($read, $expected);
+
+        $libraryB->set('isOpenOnSundays', null);
+        $this->dataLibrary->update($libraryB);
+
+        $read = $this->dataLibrary->get($libraryB->get('id'))->get('isOpenOnSundays');
+        $expected = '0';
+        $this->assertSame($read, $expected);
+    }
+
 }